:recycle: Refactor: 절대경로 안된 것들 적용
