+++ Notion_VanillaJS/src/components/index.js
@@ -1,1 +1,2 @@
 export { default as SideBar } from './SideBar/SideBar';
+export { default as Editor } from './Editor/Editor';
