+++ Notion_VanillaJS/src/store/index.js
@@ -1,0 +1,2 @@
+export { PostListStore } from './PostListStore';
+export { PostStore } from './PostStore';
